Split generic arity followup tests
