+++ docs/COMPLETION_AUDIT.md
@@ -2960,6 +2960,10 @@ and do not assume Phase 4 is ready without evidence.
   `tests/integration/frontend_diagnostics/behavior_extends.rs`, preserving
   direct, imported-parent, and transitive parent-method diagnostics while
   keeping frontend helper and generic arity diagnostics focused.
+- Generic explicit arity follow-up suppression diagnostics now live in
+  `tests/generic_diagnostics/method_type_args/arity_followups.rs`, preserving
+  function and method inference/argument follow-up checks while keeping direct
+  method type-argument diagnostics focused.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
